Reorganize SSH related commands
- Deduplicate two if blocks

- Don't run if there's no .ssh folder yet

+++ .bash_aliases
@@ -4,19 +4,14 @@ export GPG_TTY=$(tty)
 # vim loses color in tmux
 alias tmux='tmux -2'
 
-# start ssh-agent automagically
-if [ ! -S ~/.ssh/ssh_auth_sock ]; then
+# start ssh-agent and add keys automatically
+# also make sure tmux remembers keys
+if [ -d ~/.ssh ] && [ ! -S ~/.ssh/ssh_auth_sock ]; then
   eval `ssh-agent`
   ln -sf "$SSH_AUTH_SOCK" ~/.ssh/ssh_auth_sock
+  export SSH_AUTH_SOCK=~/.ssh/ssh_auth_sock
+  ssh-add -l > /dev/null || ssh-add
 fi
-export SSH_AUTH_SOCK=~/.ssh/ssh_auth_sock
-ssh-add -l > /dev/null || ssh-add
-
-# make tmux remember ssh keys
-if [ -S "$SSH_AUTH_SOCK" ] && [ ! -h "$SSH_AUTH_SOCK" ]; then
-    ln -sf "$SSH_AUTH_SOCK" ~/.ssh/ssh_auth_sock
-fi
-export SSH_AUTH_SOCK=~/.ssh/ssh_auth_sock
 
 # fzf
 [ -f ~/.fzf.bash ] && source ~/.fzf.bash
